[refactor] Refactor provider api to be internal
This commit refactors the "depprovider" api to be
internal. Only the required method is left exposed.
